requirements.txt
pyTelegramBotAPI==4.6.0
python-dotenv
pytest==8.3.4
pytest-lazy-fixture==0.6.3
pytest-asyncio==0.25.2
